pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "mirofish"
version = "0.1.0"
description = "Swarm Intelligence Prediction Engine for Prediction Markets"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "Kyeong Sun Kim"},
]
keywords = ["prediction-markets", "quant-trading", "swarm-intelligence", "llm", "polymarket"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core
    "numpy>=1.26",
    "pandas>=2.1",
    "scipy>=1.12",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "httpx>=0.27",
    "aiohttp>=3.9",
    "websockets>=12.0",

    # LLM / AI
    "anthropic>=0.40",
    "openai>=1.50",
    "langchain>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.3",
    "langgraph>=0.2",
    "chromadb>=0.5",
    "sentence-transformers>=3.0",

    # Prediction Markets
    "py-clob-client>=0.16",
    "web3>=6.15",

    # Graph / Network
    "networkx>=3.2",
    "torch-geometric>=2.5",

    # Visualization
    "plotly>=5.18",
    "dash>=2.16",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "pyvis>=0.3",

    # Calibration
    "scikit-learn>=1.4",

    # Utilities
    "rich>=13.7",
    "loguru>=0.7",
    "python-dotenv>=1.0",
    "tenacity>=8.2",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "ruff>=0.3",
    "mypy>=1.8",
    "ipykernel>=6.28",
    "jupyter>=1.0",
    "python-docx>=1.1",
]

[project.scripts]
mirofish = "src.cli:main"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "DTZ", "T20", "ICN"]
ignore = ["S101", "T201"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow",
    "integration: marks integration tests",
    "live: marks tests requiring live API access",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
